Changed substance stacks to use integers instead of floats for amount and volume, improvements to new fluid physics

+++ src/main/java/com/site21/bittermelon/common/content/items/medical/tools/SyringeItem.java
@@ -79,8 +79,8 @@ public class SyringeItem extends FluidContainerItem {
         if (entity.isShiftKeyDown()) {
             stack = consumeSubstances(stack, getLimitedTransferRate(stack), target);
         } else {
-            float spaceAvailable = getCapacity(stack) - getTotalVolume(stack);
-            float transferRate = Math.min(getTransferRate(stack), spaceAvailable);
+            int spaceAvailable = getCapacity(stack) - getTotalVolume(stack);
+            int transferRate = Math.min(getTransferRate(stack), spaceAvailable);
 
             SubstanceStack substanceStack = drawSubstanceFromEntity(target, transferRate);
             updateSubstance(stack, substanceStack);
@@ -89,7 +89,7 @@ public class SyringeItem extends FluidContainerItem {
         return stack;
     }
 
-    private @NotNull SubstanceStack drawSubstanceFromEntity(@NotNull LivingEntity entity, float transferRate) {
+    private @NotNull SubstanceStack drawSubstanceFromEntity(@NotNull LivingEntity entity, int transferRate) {
         SubstanceStack stack = new SubstanceStack(BLOOD.get(), 0);
         stack.setVolume(transferRate);
